backend/requirements.txt
Django==4.2.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
channels==4.0.0
daphne==4.2.1
channels-redis==4.1.0
psycopg2-binary==2.9.7
redis==4.6.0
